+++ ydb/mvp/oidc_proxy/mvp.cpp
@@ -225,10 +225,13 @@ void TMVP::TryGetOidcOptionsFromConfig(const YAML::Node& config) {
     }
 
     SecretName = oidc["secret_name"].as<std::string>("");
-    OpenIdConnectSettings.ClientId = oidc["client_id"].as<std::string>("yc.oauth.ydb-viewer");
+    OpenIdConnectSettings.ClientId = oidc["client_id"].as<std::string>(OpenIdConnectSettings.DEFAULT_CLIENT_ID);
     OpenIdConnectSettings.SessionServiceEndpoint = oidc["session_service_endpoint"].as<std::string>("");
     OpenIdConnectSettings.SessionServiceTokenName = oidc["session_service_token_name"].as<std::string>("");
     OpenIdConnectSettings.AuthorizationServerAddress = oidc["authorization_server_address"].as<std::string>("");
+    OpenIdConnectSettings.AuthEndpoint = oidc["auth_endpoint"].as<std::string>(OpenIdConnectSettings.DEFAULT_AUTH_ENDPOINT);
+    OpenIdConnectSettings.TokenEndpoint = oidc["token_endpoint"].as<std::string>(OpenIdConnectSettings.DEFAULT_TOKEN_ENDPOINT);
+    OpenIdConnectSettings.ExchangeEndpoint = oidc["exchange_endpoint"].as<std::string>(OpenIdConnectSettings.DEFAULT_EXCHANGE_ENDPOINT);
     Cout << "Started processing allowed_proxy_hosts..." << Endl;
     for (const std::string& host : oidc["allowed_proxy_hosts"].as<std::vector<std::string>>()) {
         Cout << host << " added to allowed_proxy_hosts" << Endl;
@@ -288,8 +291,8 @@ void TMVP::TryGetGenericOptionsFromConfig(
     if (generic["auth_profile"]) {
         auto authProfile = generic["auth_profile"].as<std::string>("y-profile");
         AuthProfile = authProfile != "n-profile"
-            ? NMVP::EAuthProfile::YProfile
-            : NMVP::EAuthProfile::NProfile;
+            ? NMVP::EAuthProfile::Yandex
+            : NMVP::EAuthProfile::Nebius;
         OpenIdConnectSettings.AuthProfile = AuthProfile;
     }
 }
