+++ ydb/mvp/oidc_proxy/oidc_client.cpp
@@ -1,13 +1,13 @@
 #include "oidc_client.h"
-#include "oidc_protected_page.h"
-#include "oidc_session_create.h"
+#include "oidc_protected_page_handler.h"
+#include "oidc_session_create_handler.h"
 
 void InitOIDC(NActors::TActorSystem& actorSystem,
               const NActors::TActorId& httpProxyId,
               const TOpenIdConnectSettings& settings) {
     actorSystem.Send(httpProxyId, new NHttp::TEvHttpProxy::TEvRegisterHandler(
                          "/auth/callback",
-                         actorSystem.Register(new NMVP::TSessionCreator(httpProxyId, settings))
+                         actorSystem.Register(new NMVP::TSessionCreateHandler(httpProxyId, settings))
                          )
                      );
