+++ ydb/mvp/oidc_proxy/oidc_protected_page_nebius.h
@@ -1,0 +1,150 @@
+#pragma once
+
+#include "oidc_protected_page.h"
+
+namespace NMVP {
+
+class THandlerSessionServiceCheckNebius : public THandlerSessionServiceCheck {
+private:
+    using TBase = THandlerSessionServiceCheck;
+
+public:
+    THandlerSessionServiceCheckNebius(const NActors::TActorId& sender,
+                                const NHttp::THttpIncomingRequestPtr& request,
+                                const NActors::TActorId& httpProxyId,
+                                const TOpenIdConnectSettings& settings)
+        : THandlerSessionServiceCheck(sender, request, httpProxyId, settings)
+    {}
+
+    void StartOidcProcess(const NActors::TActorContext& ctx) override {
+        NHttp::THeaders headers(Request->Headers);
+        LOG_DEBUG_S(ctx, EService::MVP, TStringBuilder() << "Start OIDC process");
+
+        NHttp::TCookies cookies(headers.Get("Cookie"));
+        TString sessionToken = Base64Decode(cookies.Get(CreateNameSessionCookie(Settings.ClientId)));
+
+        if (sessionToken) {
+            ExchangeSessionToken(sessionToken, ctx);
+        } else {
+            RequestAuthorizationCode(ctx);
+        }
+    }
+
+    void HandleProxy(NHttp::TEvHttpProxy::TEvHttpIncomingResponse::TPtr event, const NActors::TActorContext& ctx) {
+        NHttp::THttpOutgoingResponsePtr httpResponse;
+        if (event->Get()->Response != nullptr) {
+            NHttp::THttpIncomingResponsePtr response = event->Get()->Response;
+            LOG_DEBUG_S(ctx, EService::MVP, TStringBuilder() << "Incoming response for protected resource: " << response->Status);
+            if ((response->Status == "400" || response->Status.empty()) && RequestedPageScheme.empty()) {
+                NHttp::THttpOutgoingRequestPtr request = response->GetRequest();
+                if (!request->Secure) {
+                    LOG_DEBUG_S(ctx, EService::MVP, "Try to send request to HTTPS port");
+                    NHttp::THeadersBuilder headers {request->Headers};
+                    ForwardUserRequest(headers.Get(AUTH_HEADER_NAME), ctx, true);
+                    return;
+                }
+            }
+            NHttp::THeadersBuilder headers = GetResponseHeaders(response);
+            TStringBuf contentType = headers.Get("Content-Type").NextTok(';');
+            if (contentType == "text/html") {
+                TString newBody = FixReferenceInHtml(response->Body, response->GetRequest()->Host);
+                httpResponse = Request->CreateResponse( response->Status, response->Message, headers, newBody);
+            } else {
+                httpResponse = Request->CreateResponse( response->Status, response->Message, headers, response->Body);
+            }
+        } else {
+            httpResponse = Request->CreateResponseNotFound(NOT_FOUND_HTML_PAGE, "text/html");
+        }
+        ctx.Send(Sender, new NHttp::TEvHttpProxy::TEvHttpOutgoingResponse(httpResponse));
+        Die(ctx);
+    }
+
+    void HandleExchange(NHttp::TEvHttpProxy::TEvHttpIncomingResponse::TPtr event, const NActors::TActorContext& ctx) {
+        if (!event->Get()->Response) {
+            NHttp::THeadersBuilder ResponseHeaders;
+            ResponseHeaders.Set("Content-Type", "text/plain");
+            NHttp::THttpOutgoingResponsePtr httpResponse = Request->CreateResponse("400", "Bad Request", ResponseHeaders, event->Get()->Error);
+            ctx.Send(Sender, new NHttp::TEvHttpProxy::TEvHttpOutgoingResponse(httpResponse));
+            Die(ctx);
+        } else {
+            NHttp::THttpIncomingResponsePtr response = event->Get()->Response;
+            LOG_DEBUG_S(ctx, EService::MVP, TStringBuilder() << "Incoming response for getting access token: " << response->Status);
+            if (response->Status.StartsWith("2")) {
+                LOG_DEBUG_S(ctx, EService::MVP, "SessionService.Check(): OK");
+
+                TString iamToken;
+                static NJson::TJsonReaderConfig JsonConfig;
+                NJson::TJsonValue requestData;
+                bool success = NJson::ReadJsonTree(response->Body, &JsonConfig, &requestData);
+                if (success) {
+                    iamToken = requestData["access_token"].GetStringSafe({});
+                }
+
+                const TString authHeader = IAM_TOKEN_SCHEME + iamToken;
+                ForwardUserRequest(authHeader, ctx);
+            } else {
+                if (response->Status == "400" || response->Status == "401") {
+                    StartOidcProcess(ctx);
+                    RequestAuthorizationCode(ctx);
+                } else {
+                    NHttp::THttpOutgoingResponsePtr httpResponse;
+                    NHttp::THeadersBuilder ResponseHeaders;
+                    ResponseHeaders.Parse(response->Headers);
+                    httpResponse = Request->CreateResponse(response->Status, response->Message, ResponseHeaders, response->Body);
+                    ctx.Send(Sender, new NHttp::TEvHttpProxy::TEvHttpOutgoingResponse(httpResponse));
+                    Die(ctx);
+                }
+            }
+        }
+    }
+
+    STFUNC(StateWork) {
+        switch (ev->GetTypeRewrite()) {
+            HFunc(NHttp::TEvHttpProxy::TEvHttpIncomingResponse, HandleProxy);
+        }
+    }
+
+    STFUNC(StateExchange) {
+        switch (ev->GetTypeRewrite()) {
+            HFunc(NHttp::TEvHttpProxy::TEvHttpIncomingResponse, HandleExchange);
+        }
+    }
+
+private:
+    void ExchangeSessionToken(const TString sessionToken, const NActors::TActorContext& ctx) {
+        LOG_DEBUG_S(ctx, EService::MVP, TStringBuilder() << "Exchange session token");
+        NHttp::THttpOutgoingRequestPtr httpRequest = NHttp::THttpOutgoingRequest::CreateRequestPost(Settings.GetExchangeEndpointURL());
+        httpRequest->Set<&NHttp::THttpRequest::ContentType>("application/x-www-form-urlencoded");
+
+        NMVP::TMvpTokenator* tokenator = MVPAppData()->Tokenator;
+        TString token = "";
+        if (tokenator) {
+            token = tokenator->GetToken(Settings.SessionServiceTokenName);
+        }
+        httpRequest->Set("Authorization", token); // Bearer included
+        TStringBuilder body;
+        body << "grant_type=urn:ietf:params:oauth:grant-type:token-exchange"
+             << "&requested_token_type=urn:ietf:params:oauth:token-type:access_token"
+             << "&subject_token_type=urn:ietf:params:oauth:token-type:session_token"
+             << "&subject_token=" << sessionToken;
+        httpRequest->Set<&NHttp::THttpRequest::Body>(body);
+
+        ctx.Send(HttpProxyId, new NHttp::TEvHttpProxy::TEvHttpOutgoingRequest(httpRequest));
+
+        Become(&THandlerSessionServiceCheckNebius::StateExchange);
+    }
+
+    void RequestAuthorizationCode(const NActors::TActorContext& ctx) {
+        LOG_DEBUG_S(ctx, EService::MVP, TStringBuilder() << "Request authorization code");
+        NHttp::THttpOutgoingResponsePtr httpResponse = GetHttpOutgoingResponsePtr(TStringBuf(), Request, Settings, IsAjaxRequest);
+        ctx.Send(Sender, new NHttp::TEvHttpProxy::TEvHttpOutgoingResponse(httpResponse));
+        Die(ctx);
+    }
+
+    void ForwardUserRequest(TStringBuf authHeader, const NActors::TActorContext& ctx, bool secure = false) override {
+        THandlerSessionServiceCheck::ForwardUserRequest(authHeader, ctx, secure);
+        Become(&THandlerSessionServiceCheckNebius::StateWork);
+    }
+};
+
+}  // NMVP
