+++ ydb/mvp/oidc_proxy/oidc_session_create.h
@@ -20,13 +20,14 @@
 #include <ydb/mvp/core/mvp_tokens.h>
 #include <ydb/mvp/core/appdata.h>
 #include "openid_connect.h"
-#include "oidc_session_create_n.h"
 
 namespace NMVP {
 
 class THandlerSessionCreate : public NActors::TActorBootstrapped<THandlerSessionCreate> {
 private:
     using TBase = NActors::TActorBootstrapped<THandlerSessionCreate>;
+
+protected:
     using TSessionService = yandex::cloud::priv::oauth::v1::SessionService;
 
     const NActors::TActorId Sender;
@@ -37,9 +38,7 @@ private:
     bool IsAjaxRequest = false;
     NHttp::THeadersBuilder ResponseHeaders;
 
-    void RemoveAppliedCookie(const TString& cookieName) {
-        ResponseHeaders.Set("Set-Cookie", TStringBuilder() << cookieName << "=; Path=" << GetAuthCallbackUrl() << "; Max-Age=0");
-    }
+    virtual void RemoveAppliedCookie(const TString& cookieName) = 0;
 
     bool IsStateValid(const TString& state, const NHttp::TCookies& cookies, const NActors::TActorContext& ctx) {
         const TString cookieName {CreateNameYdbOidcCookie(Settings.ClientSecret, state)};
@@ -135,6 +134,8 @@ public:
         , Settings(settings)
         {}
 
+    virtual void RequestSessionToken(const TString&, const NActors::TActorContext&) = 0;
+
     void Bootstrap(const NActors::TActorContext& ctx) {
         NHttp::TUrlParameters urlParameters(Request->URL);
         TString code = urlParameters["code"];
@@ -144,21 +145,16 @@ public:
         NHttp::TCookies cookies(headers.Get("cookie"));
 
         if (IsStateValid(state, cookies, ctx) && !code.Empty()) {
-            NHttp::THttpOutgoingRequestPtr httpRequest = NHttp::THttpOutgoingRequest::CreateRequestPost(Settings.GetTokenEndpoint());
-            httpRequest->Set<&NHttp::THttpRequest::ContentType>("application/x-www-form-urlencoded");
-            httpRequest->Set("Authorization", "Basic " + Settings.GetAuthorizationString());
-            TStringBuilder body;
-            body << "grant_type=authorization_code&code=" << code;
-            httpRequest->Set<&NHttp::THttpRequest::Body>(body);
-            ctx.Send(HttpProxyId, new NHttp::TEvHttpProxy::TEvHttpOutgoingRequest(httpRequest));
+            RequestSessionToken(code, ctx);
         } else {
             NHttp::THttpOutgoingResponsePtr response = GetHttpOutgoingResponsePtr(TStringBuf(), Request, Settings, ResponseHeaders, IsAjaxRequest);
             ctx.Send(Sender, new NHttp::TEvHttpProxy::TEvHttpOutgoingResponse(response));
             TBase::Die(ctx);
             return;
         }
-        Become(&THandlerSessionCreate::StateWork);
     }
+
+    virtual void ProcessSessionToken(const TString& accessToken, const NActors::TActorContext&) = 0;
 
     void Handle(NHttp::TEvHttpProxy::TEvHttpIncomingResponse::TPtr event, const NActors::TActorContext& ctx) {
         NHttp::THttpOutgoingResponsePtr httpResponse;
@@ -173,31 +169,7 @@ public:
                     const NJson::TJsonValue* jsonAccessToken;
                     if (jsonValue.GetValuePointer("access_token", &jsonAccessToken)) {
                         TString accessToken = jsonAccessToken->GetStringRobust();
-                        std::unique_ptr<NYdbGrpc::TServiceConnection<TSessionService>> connection = CreateGRpcServiceConnection<TSessionService>(Settings.SessionServiceEndpoint);
-
-                        yandex::cloud::priv::oauth::v1::CreateSessionRequest requestCreate;
-                        requestCreate.Setaccess_token(accessToken);
-
-                        NMVP::TMvpTokenator* tokenator = MVPAppData()->Tokenator;
-                        TString token = "";
-                        if (tokenator) {
-                            token = tokenator->GetToken(Settings.SessionServiceTokenName);
-                        }
-                        NYdbGrpc::TCallMeta meta;
-                        SetHeader(meta, "authorization", token);
-                        meta.Timeout = TDuration::Seconds(10);
-
-                        NActors::TActorSystem* actorSystem = ctx.ActorSystem();
-                        NActors::TActorId actorId = ctx.SelfID;
-                        NYdbGrpc::TResponseCallback<yandex::cloud::priv::oauth::v1::CreateSessionResponse> responseCb =
-                            [actorId, actorSystem](NYdbGrpc::TGrpcStatus&& status, yandex::cloud::priv::oauth::v1::CreateSessionResponse&& response) -> void {
-                            if (status.Ok()) {
-                                actorSystem->Send(actorId, new TEvPrivate::TEvCreateSessionResponse(std::move(response)));
-                            } else {
-                                actorSystem->Send(actorId, new TEvPrivate::TEvErrorResponse(status));
-                            }
-                        };
-                        connection->DoRequest(requestCreate, std::move(responseCb), &yandex::cloud::priv::oauth::v1::SessionService::Stub::AsyncCreate, meta);
+                        ProcessSessionToken(accessToken, ctx);
                         return;
                     } else {
                         jsonError = "Wrong OIDC provider response: access_token not found";
@@ -217,76 +189,6 @@ public:
         }
         ctx.Send(Sender, new NHttp::TEvHttpProxy::TEvHttpOutgoingResponse(httpResponse));
         Die(ctx);
-    }
-
-    void Handle(TEvPrivate::TEvCreateSessionResponse::TPtr event, const NActors::TActorContext& ctx) {
-        LOG_DEBUG_S(ctx, EService::MVP, "SessionService.Create(): OK");
-        auto response = event->Get()->Response;
-        for (const auto& cookie : response.Getset_cookie_header()) {
-            ResponseHeaders.Set("Set-Cookie", ChangeSameSiteFieldInSessionCookie(cookie));
-        }
-        NHttp::THttpOutgoingResponsePtr httpResponse;
-        ResponseHeaders.Set("Location", RedirectUrl);
-        httpResponse = Request->CreateResponse("302", "Cookie set", ResponseHeaders);
-        ctx.Send(Sender, new NHttp::TEvHttpProxy::TEvHttpOutgoingResponse(httpResponse));
-        Die(ctx);
-    }
-
-    void Handle(TEvPrivate::TEvErrorResponse::TPtr event, const NActors::TActorContext& ctx) {
-        LOG_DEBUG_S(ctx, EService::MVP, TStringBuilder() << "SessionService.Create(): " << event->Get()->Status);
-        NHttp::THttpOutgoingResponsePtr httpResponse;
-        if (event->Get()->Status == "400") {
-            httpResponse = GetHttpOutgoingResponsePtr(event->Get()->Details, Request, Settings, ResponseHeaders, IsAjaxRequest);
-        } else {
-            ResponseHeaders.Set("Content-Type", "text/plain");
-            httpResponse = Request->CreateResponse( event->Get()->Status, event->Get()->Message, ResponseHeaders, event->Get()->Details);
-        }
-        ctx.Send(Sender, new NHttp::TEvHttpProxy::TEvHttpOutgoingResponse(httpResponse));
-        Die(ctx);
-    }
-
-    STFUNC(StateWork) {
-        switch (ev->GetTypeRewrite()) {
-            HFunc(NHttp::TEvHttpProxy::TEvHttpIncomingResponse, Handle);
-            HFunc(TEvPrivate::TEvCreateSessionResponse, Handle);
-            HFunc(TEvPrivate::TEvErrorResponse, Handle);
-        }
-    }
-};
-
-class TSessionCreator : public NActors::TActor<TSessionCreator> {
-    using TBase = NActors::TActor<TSessionCreator>;
-
-    NActors::TActorId HttpProxyId;
-    const TOpenIdConnectSettings Settings;
-
-public:
-    TSessionCreator(const NActors::TActorId& httpProxyId, const TOpenIdConnectSettings& settings)
-        : TBase(&TSessionCreator::StateWork)
-        , HttpProxyId(httpProxyId)
-        , Settings(settings)
-    {}
-
-    void Handle(NHttp::TEvHttpProxy::TEvHttpIncomingRequest::TPtr event, const NActors::TActorContext& ctx) {
-        NHttp::THttpIncomingRequestPtr request = event->Get()->Request;
-        if (request->Method == "GET") {
-            switch (Settings.AuthProfile) {
-                case NMVP::EAuthProfile::YProfile:
-                    ctx.Register(new THandlerSessionCreate(event->Sender, request, HttpProxyId, Settings));
-                    return;
-                case NMVP::EAuthProfile::NProfile:
-                    ctx.Register(new THandlerSessionCreateN(event->Sender, request, HttpProxyId, Settings));
-                    return;
-            }
-        }
-        auto response = request->CreateResponseBadRequest();
-        ctx.Send(event->Sender, new NHttp::TEvHttpProxy::TEvHttpOutgoingResponse(response));
-    }
-
-    STFUNC(StateWork) {
-        switch (ev->GetTypeRewrite()) {
-            HFunc(NHttp::TEvHttpProxy::TEvHttpIncomingRequest, Handle);
-        }
     }
 };
